backend/requirements.txt
fastapi==0.121.0
pydantic-settings==2.3.4
python-multipart==0.0.9
uvicorn[standard]==0.38.0
google-cloud-firestore==2.21.0